plaid/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
idna==3.4
nulltype==2.3.1
plaid-python==10.0.0
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12